+++ app/javascript/packs/controllers/modal_controller.js
@@ -11,8 +11,8 @@ export default class extends Controller {
     this.modal(event).classList.remove("modal-opened");
   }
 
-  modal(event) {
-    if (event.currentTarget.dataset.selector){      
+  modal(event) {    
+    if (event.currentTarget.dataset.selector){
       return document.getElementById(event.currentTarget.dataset.selector);
     }
     return this.modalWindowTarget;