+++ app/models/player.rb
@@ -13,10 +13,14 @@ class Player < ApplicationRecord
   accepts_nested_attributes_for :socials, allow_destroy: true, reject_if: proc { |att| att[:link].blank? }
   has_many :comments
   accepts_nested_attributes_for :comments, allow_destroy: true, reject_if: proc { |att| att[:content].blank? }
-
+  
+  has_and_belongs_to_many :favorited_by_users, class_name: "User", join_table: "players_users"
+  has_many :player_spots
+  has_many :roster_spots, through: :player_spots
   belongs_to :user, foreign_key: "updated_by_user_id"
   validates :name, :nationality, presence: true
-  
+
+  default_scope { includes({ comments: :user}, :user, :position, :status, :priority, :profiles, :seasons) }
 
   scope :with_height_between, -> (min, max) { where(height_eu: min..max) }
   scope :with_weight_between, -> (min, max) { where(weight_eu: min..max) }
@@ -70,5 +74,10 @@ class Player < ApplicationRecord
     self.height_eu = 0 if self.height_eu.nil?
     self.weight_eu = 0 if self.weight_eu.nil?
     self.salary_real = 0 if self.salary_real.nil?
+  end
+
+  def touch_by(user)
+    self.touch
+    self.update_column(:updated_by_user_id, user.id)
   end
 end