+++ app/models/roster_spot.rb
@@ -1,0 +1,6 @@
+class RosterSpot < ApplicationRecord
+  belongs_to :roster
+  belongs_to :position
+  has_many :player_spots
+  has_many :players, through: :player_spots
+end